fix(crd): Change ObservedGeneration from int64 to int for OpenAPI compatibility
OpenAPI 3.0 spec only supports 'int32' format, not 'int64'. Using int64
caused a validation warning:

  Warning: unrecognized format 'int64'

Changed ObservedGeneration field from int64 to int (which maps to int32
in the generated OpenAPI schema). This is safe as generation counters
will never exceed int32 range in practice.

Note: One warning remains from metav1.Condition.observedGeneration which
is an upstream Kubernetes type that cannot be modified.

+++ api/v1alpha1/resultsexporter_types.go
@@ -215,7 +215,7 @@ type ResultsExporterStatus struct {
 
 	// Observed generation
 	// +optional
-	ObservedGeneration int64 `json:"observedGeneration,omitempty"`
+	ObservedGeneration int `json:"observedGeneration,omitempty"`
 }
 
 // ExportedResourceCounts tracks how many resources have been exported
